Add an API between backend and database classes.
