+++ lsif/src/dump-manager/settings.ts
@@ -5,3 +5,16 @@ export const HTTP_PORT = readEnvInt('HTTP_PORT', 3187)
 
 /** Where on the file system to store LSIF files. */
 export const STORAGE_ROOT = process.env.LSIF_STORAGE_ROOT || 'lsif-storage'
+
+/**
+ * The number of SQLite connections that can be opened at once. This
+ * value may be exceeded for a short period if many handles are held
+ * at once.
+ */
+export const CONNECTION_CACHE_CAPACITY = readEnvInt('CONNECTION_CACHE_CAPACITY', 100)
+
+/** The maximum number of documents that can be held in memory at once. */
+export const DOCUMENT_CACHE_CAPACITY = readEnvInt('DOCUMENT_CACHE_CAPACITY', 1024 * 1024 * 1024)
+
+/** The maximum number of result chunks that can be held in memory at once. */
+export const RESULT_CHUNK_CACHE_CAPACITY = readEnvInt('RESULT_CHUNK_CACHE_CAPACITY', 1024 * 1024 * 1024)
