+++ lsif/src/server/settings.ts
@@ -3,6 +3,9 @@ import { readEnvInt } from '../shared/settings'
 /** Which port to run the LSIF server on. Defaults to 3186. */
 export const HTTP_PORT = readEnvInt('HTTP_PORT', 3186)
 
+/** HTTP address for internal LSIF dump manager server. */
+export const LSIF_DUMP_MANAGER_URL = process.env.LSIF_DUMP_MANAGER_URL || 'http://lsif-dump-manager'
+
 /** Where on the file system to store LSIF files. */
 export const STORAGE_ROOT = process.env.LSIF_STORAGE_ROOT || 'lsif-storage'
 
@@ -32,19 +35,6 @@ export const DEFAULT_UPLOAD_PAGE_SIZE = readEnvInt('DEFAULT_UPLOAD_PAGE_SIZE', 5
 
 /** The default page size for the dumps endpoint. */
 export const DEFAULT_DUMP_PAGE_SIZE = readEnvInt('DEFAULT_DUMP_PAGE_SIZE', 50)
-
-/**
- * The number of SQLite connections that can be opened at once. This
- * value may be exceeded for a short period if many handles are held
- * at once.
- */
-export const CONNECTION_CACHE_CAPACITY = readEnvInt('CONNECTION_CACHE_CAPACITY', 100)
-
-/** The maximum number of documents that can be held in memory at once. */
-export const DOCUMENT_CACHE_CAPACITY = readEnvInt('DOCUMENT_CACHE_CAPACITY', 1024 * 1024 * 1024)
-
-/** The maximum number of result chunks that can be held in memory at once. */
-export const RESULT_CHUNK_CACHE_CAPACITY = readEnvInt('RESULT_CHUNK_CACHE_CAPACITY', 1024 * 1024 * 1024)
 
 /** The maximum age (in seconds) that an upload (completed or queued) will remain in Postgres. */
 export const UPLOAD_MAX_AGE = readEnvInt('UPLOAD_UPLOAD_AGE', 60 * 60 * 24 * 7)
